Updated State Machine to try to work with the DataFile_Handler
Shit don’t work.
